Combat speed is now fractional; there is a buffer which saves up fractional combat speed for a speed boost. In stock this means that odd speed ships will only move faster every other turn.

+++ FrEee/Game/Interfaces/ICombatant.cs
@@ -18,7 +18,7 @@ namespace FrEee.Game.Interfaces
 		/// <summary>
 		/// How fast can this combatant move in combat?
 		/// </summary>
-		int CombatSpeed { get; }
+		double CombatSpeed { get; }
 
 		/// <summary>
 		/// Is this combatant still alive or is it destroyed/glassed?
